% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asiaverse-package.R, R/asiaverse.R
\docType{package}
\name{Asiaverse}
\alias{Asiaverse}
\alias{Asiaverse-package}
\title{Asiaverse: A Metapackage for Asian Countries RESTful APIs and Curated Datasets}
\usage{
Asiaverse()
}
\value{
Invisibly returns the names of the loaded packages.
}
\description{
{
This metapackage brings together a comprehensive collection of R packages providing access to APIs functions and curated datasets from China, Japan, South Korea, India and Indonesia. It includes real-time and historical data through public RESTful APIs (Nager.Date, World Bank API, REST Countries API) and extensive curated collections of open datasets covering economics, demographics, public health, environmental data, natural disasters, political indicators, and social metrics.
}

This function displays a formatted list of the API packages included
in the Asiaverse metapackage and their respective versions.
}
\details{
Asiaverse: A Metapackage for Asian Countries RESTful APIs and Curated Datasets

\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

A Metapackage for Asian Countries RESTful APIs and Curated Datasets.
}
\seealso{
{
Useful links:
\itemize{
  \item \url{https://github.com/lightbluetitan/asiaverse}
}
}
}
\author{
{
\strong{Maintainer}: Renzo Caceres Rossi \email{arenzocaceresrossi@gmail.com}
}
}
